refactor(blackjack/): 파일의 상위 패키지 변경
BlackJack.java 상위 패키지 변경
Roles.java 상위 패키지 변경

BREAKING CHANGE: Roles.java의 상위 패키지를 blackjack에서 role로
변경. BlackJack.java의 상위 패키지를 blackjack에서 domain으로 변경
