+++ src/main/java/blackjack/BlackJackApplication.java
@@ -1,5 +1,7 @@
 package blackjack;
 
+import blackjack.domain.BlackJack;
+
 public class BlackJackApplication {
 
 	public static void main(String[] args) {
